fix clang warnings
git-svn-id: http://root.cern.ch/svn/root/trunk@36396 27541ba8-7e3a-0410-8455-c3a389f83636

+++ bindings/pyroot/src/Converters.cxx
@@ -680,7 +680,7 @@ Bool_t PyROOT::T##name##ArrayConverter::ToMemory( PyObject* value, void* address
          PyErr_SetString( PyExc_ValueError, "buffer too large for value" );  \
          return kFALSE;                                                      \
       }                                                                      \
-      memcpy( *(type**)address, buf, 0 < buflen ? buflen : sizeof(type) );   \
+      memcpy( *(type**)address, buf, 0 < buflen ? ((size_t) buflen) : sizeof(type) );\
    } else                                                                    \
       *(type**)address = (type*)buf;                                         \
    return kTRUE;                                                             \
